requirements.txt
attrs==24.3.0
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-core==1.5.0
azure-storage-blob==12.24.0
black==24.10.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
et_xmlfile==2.0.0
Flask==3.1.0
fsspec==2024.12.0
h11==0.14.0
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
lxml==5.3.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
msal==1.31.1
msal-extensions==1.2.0
mypy-extensions==1.0.0
numpy==2.2.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
portalocker==2.10.1
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytweening==1.2.0
pytz==2024.2
pywin32==308
requests==2.32.3
screeninfo==0.8.1
secure-smtplib==0.1.1
selenium==4.27.1
selenium-stealth==1.0.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
XlsxWriter==3.2.0
